+++ @types/yata/process.d.ts
@@ -5,5 +5,6 @@ declare namespace NodeJS {
     REACT_APP_TWITCH_CLIENT_ID: string
     REACT_APP_TWITCH_REDIRECT_URI: string
     REACT_APP_YOUTUBE_API_KEY: string
+    REACT_APP_IMGUR_CLIENT_ID: string
   }
 }